Extract Keycloak auth/session ops into KeycloakAuthClient.
Keeps IdentityClient facade intact while isolating login, logout, OTP-ignore verify, and session close for safer incremental refactor (#91).
